refactor(visualization): move gameObjectsParser util into navBar feature
Sole consumer is navBar/services/readFiles.ts; relocate the whole
gameObjectsParser module (importer, validator, schema, mocks, specs)
into features/navBar/util/. Verified by dependency-cruiser (0 errors)
and jest.
