+++ stories/cube-slider-daisyui.md
@@ -1,0 +1,13 @@
+# Restyle Cube Slider with daisyUI
+
+**As a** user of the visualization
+**I want** the cube slider to use daisyUI styling
+**So that** it looks and behaves consistently with the rest of the application
+
+## Acceptance Criteria
+- [ ] The cube slider is rebuilt using daisyUI (Tailwind) components instead of its current styling
+- [ ] No Angular Material or custom one-off styles remain for this slider
+- [ ] The slider matches the look and feel (colors, spacing, states) of other daisyUI controls in the app
+- [ ] Existing behavior is preserved: value range, step, live update of the map, and bound state stay unchanged
+- [ ] Hover, focus, active, and disabled states are styled and accessible (keyboard operable)
+- [ ] Tests/snapshots are updated to reflect the new markup
