+++ visualization/app/codeCharta/features/navBar/util/gameObjectsParser/gameObjectsImporter.spec.ts
@@ -1,5 +1,5 @@
 import { TEST_GAMEOBJECTS_FILE } from "./gameObjectsMocks"
-import { clone } from "../clone"
+import { clone } from "../../../../util/clone"
 import { parseGameObjectsFile } from "./gameObjectsImporter"
 
 describe("GameObjectsImporter", () => {
