+++ visualization/app/codeCharta/features/navBar/util/gameObjectsParser/gameObjectsMocks.ts
