+++ visualization/app/codeCharta/features/navBar/util/gameObjectsParser/gameObjectsSchema.json
