+++ visualization/app/codeCharta/features/navBar/util/gameObjectsParser/gameObjectsValidator.ts
